requirements.txt
python-telegram-bot==21.6
SQLAlchemy==2.0.25
aiosqlite==0.19.0
asyncpg==0.29.0
imagehash==4.3.1
Pillow==10.2.0
cachetools==5.3.2
python-dotenv==1.0.0
aiofiles==23.2.1
alembic==1.14.0
pytest==7.4.4
pytest-asyncio==0.23.3
